MBS-9655 remove graphite writer (not used)
