+++ subprojects/gradle/test-summary/build.gradle.kts
@@ -11,7 +11,6 @@ dependencies {
     implementation(project(":gradle:ci-logger"))
     implementation(project(":gradle:kotlin-dsl-support"))
     implementation(project(":gradle:slack"))
-    implementation(project(":gradle:statsd-config"))
     implementation(Dependencies.funktionaleTry)
     implementation(Dependencies.okhttp)
