+++ subprojects/gradle/test-summary/src/main/kotlin/com/avito/test/summary/FlakyReportTask.kt
@@ -92,8 +92,7 @@ abstract class FlakyReportTask : DefaultTask() {
             reportViewer = reportViewer,
             buildUrl = buildUrl,
             currentBranch = currentBranch,
-            reportCoordinates = reportCoordinates,
-            logger = logger
+            reportCoordinates = reportCoordinates
         )
     }
 }
